+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 * Add `EveOnline::ESI::UniverseSystemKills`
 * Add `EveOnline::ESI::UniverseStructures`
 * Add `EveOnline::ESI::Models::StargateDestination`
+* Add `EveOnline::ESI::Models::Stargate`
 
 ## v0.19.0